Introduce UIUtils package.
